+++ lib/rpcBase/index.ts
@@ -33,7 +33,6 @@ export type TResponse<T = unknown, R = boolean> = {
 export type TOptions = {
   protocol?: "http" | "https";
   rpcURI?: string;
-  rpcLoadfileURI?: string;
 };
 
 const axiosInstance = axios.create({
@@ -47,10 +46,9 @@ const axiosInstance = axios.create({
 export class RPCBase {
   readonly baseURL;
   readonly rpcURI;
-  readonly rpcLoadfileURL;
   _id = 0;
   _session = "";
-  _cookies = "";
+  _username = "";
 
   Global = Global.bind(this)();
 
@@ -60,8 +58,6 @@ export class RPCBase {
   constructor(address: string, options: TOptions = {}) {
     this.baseURL = `${options.protocol ?? "http"}://${address}`;
     this.rpcURI = options.rpcURI ?? "/RPC2";
-    this.rpcLoadfileURL =
-      this.baseURL + (options.rpcLoadfileURI ?? "/RPC_Loadfile");
   }
 
   _nextID(): number {
@@ -69,20 +65,16 @@ export class RPCBase {
   }
 
   _setSession(username: string, session: string) {
-    this._cookies = `username=${username}; WebClientSessionID=${session}; DWebClientSessionID=${session}; DhWebClientSessionID=${session}`;
+    this._username = username;
     this._session = session;
+  }
+
+  getSessionUsername(): string {
+    return this._username;
   }
 
   getSession(): string {
     return this._session;
-  }
-
-  getCookies(): string {
-    return this._cookies;
-  }
-
-  getRPCLoadfileURL(filePath: string): string {
-    return `${this.rpcLoadfileURL}${filePath}`;
   }
 
   /**